Bill Splitter Application
This is another simple but slightly more complex application that allows a user to add their friends to the list and split the bill of anything they may need to split. Another application i recently felt in need of myself and felt i t is another good project to keep practicing. This application is not finished and is still in need of some style changes as well as a currently known bug when switching friends while having details in the in the split bill form the fields will not clear. I will fix this and add finishing touches to it soon.